Create week 3
GreedyMotifSearch is fast. In fact it trades speed for accuracy. When it try to find the (4, 1)-motif 'ACGT' implanted in the following strings DNA:

ttACCTtaac
gATGTctgtc
acgGCGTtag
ccctaACGAg
cgtcagAGGT

If we assume that the algorithm has already correctly chosen the implanted 4-mer 'ACCT' from the first string DNA and constructed the corresponding Profile:

A: 1 0 0 0
C: 0 1 1 0
G: 0 0 0 0
T: 0 0 0 1

This algorithm is now ready to search for a Profile-most probable 4-mer in the second string. However, according to this profile matrix, the probability of every 4-mer but 'ACCT' is zero! Thus, unless 'ACCT' is present in every string DNA, there is little chance that GreedyMotifSearch will find the implanted motif. Zeroes in the profile matrix are not just a minor annoyance but rather a persistent problem.