Update week6_SquaredErrorDistortion.py
Center placement varies in different clustering problem formulations.In the k-Center Clustering Problem, a cluster’s center is chosen so that the maximum distance between the center and any point in the cluster is minimized. As a result, the position of the center can be greatly influenced by outliers. In the k-Means Clustering Problem, the outlier’s influence over the placement of the center is much smaller. This is preferable when analyzing biological datasets, in which outliers often correspond to erroneous data.